Add cartopy to list of dependencies.
Move functions related to maps to own submodule.
