+++ typhon/plots/__init__.py
@@ -7,6 +7,7 @@ from typhon.plots import cm  # noqa
 from typhon.plots.colors import *  # noqa
 from typhon.plots.common import *  # noqa
 from typhon.plots.formatter import *  # noqa
+from typhon.plots.maps import *  # noqa
 from typhon.plots.plots import *  # noqa
 from typhon.plots.arts_lookup import *  # noqa
